Dev-build filter migration
