+++ dev-build/conf/dev-build.application.conf
@@ -49,6 +49,7 @@ play {
   http {
     requestHandler: "http.DevBuildParametersHttpRequestHandler"
     errorHandler: "http.CorsHttpErrorHandler"
+    filters: "http.DevFilters"
   }
 }
